Split files.
In addition, include sdl3_image and create a texture.
