+++ include/app.h
@@ -5,6 +5,8 @@
 
 typedef struct {
 	SDL_Window *window;
+	SDL_Renderer *renderer;
+	SDL_Texture *player_texture;
 } app_t;
 
 #endif
