+++ src/main.c
@@ -1,11 +1,17 @@
 #define SDL_MAIN_USE_CALLBACKS
 #include <SDL3/SDL_main.h>
 #include <SDL3/SDL.h>
+#include <SDL3_image/SDL_Image.h>
 
 #include <stdlib.h>
 #include <stdio.h>
 
 #include "app.h"
+
+void log_error(char* prefix)
+{
+	SDL_Log("%s: %s", prefix, SDL_GetError());
+}
 
 SDL_AppResult SDL_AppInit(void **appstate, int argc, char *argv[])
 {
@@ -24,21 +30,22 @@ SDL_AppResult SDL_AppInit(void **appstate, int argc, char *argv[])
 			600,
 			0);
 
-	return SDL_APP_CONTINUE;
-}
-
-SDL_AppResult SDL_AppIterate(void *appstate)
-{
-	return SDL_APP_CONTINUE;
-}
-
-SDL_AppResult SDL_AppEvent(void *appstate, SDL_Event *event)
-{
-	switch (event->type)
+	if (!state->window) 
 	{
-		case SDL_EVENT_QUIT:
-			return SDL_APP_SUCCESS;
+		log_error("Error initializing window"); 
+		return SDL_APP_FAILURE; 
+	}	
+
+	state->renderer = SDL_CreateRenderer(state->window, NULL);
+
+	if (!state->renderer)
+	{
+		log_error("Error loading renderer");
+		return SDL_APP_FAILURE;
 	}
+
+ 
+	state->player_texture = IMG_LoadTexture(state->renderer, "./char_spritesheet.png");	
 
 	return SDL_APP_CONTINUE;
 }
